backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
SQLAlchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
redis>=5.0
minio>=7.2
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt>=4.1
